refactor: move computed status query logic from managers to selectors
Per HackSoft style guide: "Business logic should not live in Custom
managers or querysets."

Move all query-building functions encoding the status priority chain
(status_filter_q, computed_status_case, in_turnaround_filter_q,
has_*_exists helpers, shelter_*_status_count_subquery) from
shelters/managers.py into a new shelters/selectors/computed_status.py.

Keep only thin Manager/QuerySet wrappers in managers.py that delegate
to the selector functions via imports.

This completes the HackSoft compliance for this layer:
- Models: thin (fields, simple clean(), shelter @property)
- Services: write-side logic (timestamp auto-set, complex validation)
- Selectors: read-side + query-building logic (computed_status, filters)
